#200 Fix: Account Deletion

+++ app/Actions/Jetstream/DeleteUser.php
@@ -4,6 +4,7 @@ namespace App\Actions\Jetstream;
 
 use Auth;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Schema;
 use Illuminate\Validation\ValidationException;
 use Laravel\Jetstream\Contracts\DeletesTeams;
 use Laravel\Jetstream\Contracts\DeletesUsers;
@@ -39,7 +40,7 @@ class DeleteUser implements DeletesUsers
         $pkg_status = \App\Models\User::whereRelation('roles', 'name', 'user')
             ->where('id', Auth::user()->id)
             ->whereDoesntHave('purchasedPackages')
-            ->whereDoesntHave('transactions', fn($q) => $q->where('status', 'PENDING'))
+            ->whereDoesntHave('transactions', fn($q) => $q->whereIn('status', ['PENDING', 'PAID']))
             ->doesntExist();
         if ($pkg_status) {
 //            session()->flash('error', 'You are not allowed to delete this account! Contact your administrator for further instructions.');
@@ -47,12 +48,23 @@ class DeleteUser implements DeletesUsers
                 'account' => [__('You are not allowed to delete this account! Contact your administrator for further instructions.')],
             ]);
         }
+
+        if ($user->descendants()->count() > 0) {
+//            session()->flash('error', 'You are not allowed to delete this account! Contact your administrator for further instructions.');
+            throw ValidationException::withMessages([
+                'account' => [__('You are not allowed to delete this account! Because you have referral users under your account. Please Contact your administrator for further instructions.')],
+            ]);
+        }
+
         DB::transaction(function () use ($user) {
+            Schema::disableForeignKeyConstraints();
             $this->deleteTeams($user);
             $user->deleteProfilePhoto();
             $user->tokens->each->delete();
+            $user->transactions()->delete();
             $user->profile()->forceDelete();
             $user->forceDelete();
+            Schema::enableForeignKeyConstraints();
 //            $user->delete();
         });
     }
